Optimization to load only needed .pc files
Currently pkg-config scans all .pc files from the search path during
initialization. That makes some of the code simpler and works fine when
there are not many .pc files on the system. When there are a lot of .pc
files, then this represents a lot of wasted effort.

Rework the package gathering so that it happens as needed. To support
the --list-all mode, the scanning at initialization can still be done.

https://bugs.freedesktop.org/show_bug.cgi?id=98215