parse: Don't emit unknown keyword warning for Libs.private
Requires.private and Libs.private are skipped unless --static is in use.
However, the parser was checking for Libs.private and this option in a
single conditional unlike Requires.private. This was making the check
fall through to the else and emit an "Unknown keyword" debug message
when --static was not used. This was harmless but not true.

https://bugs.freedesktop.org/show_bug.cgi?id=99770